database migration

+++ backend-laravel/routes/api.php
@@ -14,25 +14,37 @@ Route::prefix('auth')->group(function () {
 });
 
 Route::middleware('auth:api')->group(function () {
-    Route::get('users', 'Api\UserController@index');
-    Route::get('roles', 'Api\RoleController@index');
-    Route::get('permissions', 'Api\PermissionController@index');
+    Route::middleware(['module_enabled:users', 'permission:users.view'])->get('users', 'Api\UserController@index');
+    Route::middleware(['module_enabled:roles', 'permission:roles.view'])->get('roles', 'Api\RoleController@index');
+    Route::middleware(['module_enabled:permissions', 'permission:permissions.view'])->get('permissions', 'Api\PermissionController@index');
 
-    Route::middleware('role:admin')->group(function () {
-        Route::post('users', 'Api\UserController@store');
-        Route::put('users/{user}', 'Api\UserController@update');
-        Route::delete('users/{user}', 'Api\UserController@destroy');
-        Route::post('users/{user}/roles', 'Api\UserController@assignRoles');
-    });
+    Route::middleware(['module_enabled:users', 'permission:users.create'])->post('users', 'Api\UserController@store');
+    Route::middleware(['module_enabled:users', 'permission:users.edit'])->put('users/{user}', 'Api\UserController@update');
+    Route::middleware(['module_enabled:users', 'permission:users.delete'])->delete('users/{user}', 'Api\UserController@destroy');
+    Route::middleware(['module_enabled:users', 'permission:users.edit'])->post('users/{user}/roles', 'Api\UserController@assignRoles');
 
-    Route::middleware('role:super-admin')->group(function () {
-        Route::post('roles', 'Api\RoleController@store');
-        Route::put('roles/{role}', 'Api\RoleController@update');
-        Route::delete('roles/{role}', 'Api\RoleController@destroy');
-        Route::post('roles/{role}/permissions', 'Api\RoleController@assignPermissions');
+    Route::middleware(['module_enabled:roles', 'permission:roles.create'])->post('roles', 'Api\RoleController@store');
+    Route::middleware(['module_enabled:roles', 'permission:roles.edit'])->put('roles/{role}', 'Api\RoleController@update');
+    Route::middleware(['module_enabled:roles', 'permission:roles.delete'])->delete('roles/{role}', 'Api\RoleController@destroy');
+    Route::middleware(['module_enabled:roles', 'permission:roles.edit'])->post('roles/{role}/permissions', 'Api\RoleController@assignPermissions');
 
-        Route::post('permissions', 'Api\PermissionController@store');
-        Route::put('permissions/{permission}', 'Api\PermissionController@update');
-        Route::delete('permissions/{permission}', 'Api\PermissionController@destroy');
-    });
+    Route::middleware(['module_enabled:permissions', 'permission:permissions.create'])->post('permissions', 'Api\PermissionController@store');
+    Route::middleware(['module_enabled:permissions', 'permission:permissions.edit'])->put('permissions/{permission}', 'Api\PermissionController@update');
+    Route::middleware(['module_enabled:permissions', 'permission:permissions.delete'])->delete('permissions/{permission}', 'Api\PermissionController@destroy');
+
+    Route::middleware(['module_enabled:institutions', 'permission:institutions.view'])->get('institutions', 'Api\InstitutionController@index');
+    Route::middleware(['module_enabled:institutions', 'permission:institutions.create'])->post('institutions', 'Api\InstitutionController@store');
+    Route::middleware(['module_enabled:institutions', 'permission:institutions.view'])->get('institutions/{id}', 'Api\InstitutionController@show');
+    Route::middleware(['module_enabled:institutions', 'permission:institutions.edit'])->put('institutions/{id}', 'Api\InstitutionController@update');
+    Route::middleware(['module_enabled:institutions', 'permission:institutions.delete'])->delete('institutions/{id}', 'Api\InstitutionController@destroy');
+    Route::middleware(['module_enabled:institutions', 'permission:institutions.settings'])->get('institutions/{id}/settings', 'Api\InstitutionController@getSettings');
+    Route::middleware(['module_enabled:institutions', 'permission:institutions.settings'])->put('institutions/{id}/settings', 'Api\InstitutionController@updateSettings');
+
+    Route::middleware(['module_enabled:institutions', 'permission:institutions.edit'])->post('institutions/{id}/upload-logo', 'Api\InstitutionController@uploadLogo');
+    Route::middleware(['module_enabled:institutions', 'permission:institutions.edit'])->post('institutions/{id}/upload-letterhead', 'Api\InstitutionController@uploadLetterhead');
+    Route::middleware(['module_enabled:institutions', 'permission:institutions.edit'])->post('institutions/{id}/upload-signature', 'Api\InstitutionController@uploadSignature');
+    Route::middleware(['module_enabled:institutions', 'permission:institutions.edit'])->post('institutions/{id}/upload-stamp', 'Api\InstitutionController@uploadStamp');
+    Route::middleware('permission:modules.view')->get('modules', 'Api\ModuleController@index');
+    Route::middleware('permission:modules.manage')->get('institutions/{institution}/modules', 'Api\InstitutionModuleController@show');
+    Route::middleware('permission:modules.manage')->put('institutions/{institution}/modules', 'Api\InstitutionModuleController@update');
 });
